requirements.txt
configparser==4.0.2
python-twitter==3.5